Moved server functions from genericio to their own function and updated method signatures
